重构查询框功能设计 新增full查询
