+++ src/main/java/org/fofaviewer/bean/FofaBean.java
@@ -8,14 +8,17 @@ public class FofaBean {
     public String page = "1";
     public final int max = 10000;
     public String size = "1000";
-    public static final String TIP_API = Locale.getDefault()==Locale.CHINA ? "https://api.fofa.so/v1/search/tip?q=" : "https://api.soall.org/v1/search/tip?q=";
-    public static final String API = "https://fofa.so/api/v1/search/all";
+    public String API = "https://fofa.info";
+    public final String path = "/api/v1/search/all";
+    public static final String TIP_API = "https://api.fofa.info/v1/search/tip?q=";
+    //public static final String TIP_API = Locale.getDefault()==Locale.CHINA ? "https://api.fofa.so/v1/search/tip?q="
     public static final String fields = "host,title,ip,domain,port,protocol,server,cert";
     public static final String fid = "fid";
 
-    public FofaBean(String m, String k) {
+    public FofaBean(String m, String k, String a) {
         this.email = m;
         this.key = k;
+        this.API = a + this.path;
     }
 
     public String getSize() {
@@ -26,9 +29,10 @@ public class FofaBean {
         this.size = size;
     }
 
-    public String getParam(String page, boolean hasFid) {
+    public String getParam(String page, boolean hasFid, boolean isAll) {
         String fields = hasFid ? FofaBean.fields + "," +fid : FofaBean.fields;
-        StringBuilder builder = new StringBuilder(API).append("?email=").append(email).append("&key=").append(key).append("&page=");
+        String all = isAll ? "&full=true" : "";
+        StringBuilder builder = new StringBuilder(API).append("?email=").append(email).append("&key=").append(key).append(all).append("&page=");
         if(page != null) {
             return builder.append(page).append("&size=").append(size).append("&fields=").append(fields).append("&qbase64=").toString();
         }else{
