+++ src/main/java/org/fofaviewer/controls/CloseableTabPane.java
@@ -1,7 +1,6 @@
 package org.fofaviewer.controls;
 
 import javafx.collections.ListChangeListener;
-import javafx.collections.ObservableList;
 import javafx.geometry.Insets;
 import javafx.geometry.Pos;
 import javafx.scene.control.*;
@@ -10,12 +9,16 @@ import javafx.scene.layout.StackPane;
 import org.controlsfx.control.StatusBar;
 import org.fofaviewer.bean.TabDataBean;
 import org.fofaviewer.utils.ResourceBundleUtil;
+
+import java.util.ArrayList;
 import java.util.HashMap;
 
 public class CloseableTabPane extends BorderPane {
     private final String homepage = ResourceBundleUtil.getResource().getString("HOMEPAGE");
     private final TabPane tabPane;
+    // tab页的数据存储，主要是
     private HashMap<Tab, TabDataBean> dataMap;
+    // tab页下方的状态栏
     private HashMap<Tab, StatusBar> barMap;
 
     public CloseableTabPane() {
@@ -40,45 +43,41 @@ public class CloseableTabPane extends BorderPane {
         //关闭选中的Tab
         MenuItem closeSelected = new MenuItem(ResourceBundleUtil.getResource().getString("CLOSE_CHOOSE"));
         closeSelected.setOnAction(e->{
-            for (Tab tab:tabPane.getTabs()){
-                if(tab.getText().equals(homepage)){
-                    continue;
-                }
-                if(tab.selectedProperty().getValue()){
-                    tabPane.getTabs().remove(tab);
-                    dataMap.remove(tab);
-                    break;
-                }
+            Tab tab = this.tabPane.getSelectionModel().getSelectedItem();
+            if(!tab.getText().equals(homepage)){
+                tabPane.getTabs().remove(tab);
+                dataMap.remove(tab);
             }
         });
-        menuButton.getItems().add(closeSelected);
         //关闭除选中的其他Tab
         MenuItem closeOthers = new MenuItem(ResourceBundleUtil.getResource().getString("CLOSE_OTHERS"));
-        menuButton.getItems().add(closeOthers);
         closeOthers.setOnAction(e->{
-            for (Tab tab:tabPane.getTabs()){
-                if(tab.getText().equals(homepage)){ // 阻止关闭起始页
-                    continue;
-                }
-                if(tab.selectedProperty().getValue()){
-                    tabPane.getTabs().clear();
-                    tabPane.getTabs().add(tab);
-                    TabDataBean tmp = dataMap.get(tab);
-                    dataMap.clear();
-                    dataMap.put(tab, tmp);
-                    break;
-                }
-            }
+            if(tabPane.getTabs().size()==1) return;
+            Tab tab = this.tabPane.getSelectionModel().getSelectedItem();
+            Tab home = this.getTab(homepage);
+            if(tab.equals(home)) return;
+            tabPane.getTabs().clear();
+            tabPane.getTabs().addAll(home, tab);
+            TabDataBean tmp = dataMap.get(tab);
+            dataMap.clear();
+            dataMap.put(tab, tmp);
         });
         //关闭所有的Tab
         MenuItem closeAll = new MenuItem(ResourceBundleUtil.getResource().getString("CLOSE_ALL"));
         closeAll.setOnAction(e->{
+            if(tabPane.getTabs().size()==1) return;
             Tab tab = this.getTab(homepage);
             tabPane.getTabs().clear();
             dataMap.clear();
-            this.addTab(tab, null);
+            tabPane.getTabs().add(tab);
         });
-        menuButton.getItems().add(closeAll);
+        MenuItem reload = new MenuItem(ResourceBundleUtil.getResource().getString("RELOAD_TAB"));
+        reload.setOnAction(e->{
+            Tab tab = this.tabPane.getSelectionModel().getSelectedItem();
+            TabDataBean bean = this.getTabDataBean(tab);
+
+        });
+        menuButton.getItems().addAll(closeOthers, closeSelected, closeAll, reload);
         sp.getChildren().add(menuButton);
         super.setCenter(sp);
         Tab tab = new Tab(homepage);
@@ -93,13 +92,6 @@ public class CloseableTabPane extends BorderPane {
         if(list != null){
             dataMap.put(tab, list);
         }
-    }
-    //添加Tab集合
-    public void addAll(ObservableList<Tab> tabs){
-        for (Tab tab:tabs){
-            tab.setClosable(true);
-        }
-        tabPane.getTabs().addAll(tabs);
     }
 
     public void addBar(Tab tab, StatusBar bar){
@@ -126,6 +118,10 @@ public class CloseableTabPane extends BorderPane {
             }
         }
         return null;
+    }
+
+    public ArrayList<Tab> getTabs(){
+        return new ArrayList<>(this.tabPane.getTabs());
     }
 
     public TabDataBean getTabDataBean(Tab tab){
