+++ src/main/java/org/fofaviewer/main/MainApp.java
@@ -6,6 +6,7 @@ import javafx.scene.Scene;
 import javafx.stage.Stage;
 
 public class MainApp extends Application {
+    private String version = "1.1.5";
 
     @Override
     public void init() throws Exception {
@@ -17,7 +18,7 @@ public class MainApp extends Application {
         FXMLLoader loader = new FXMLLoader(getClass().getResource("/fofa_viewer.fxml"));
         Scene scene = new Scene(loader.load());
         stage.setScene(scene);
-        stage.setTitle("Fofa_Viewer v1.1.4 By f1ashine@WgpSec");
+        stage.setTitle("Fofa_Viewer v" + this.version+ " By f1ashine@WgpSec");
         stage.show();
         stage.setMinWidth(1100);
         stage.setMinHeight(800);
