+++ src/main/java/org/fofaviewer/request/MainControllerCallback.java
@@ -1,0 +1,17 @@
+package org.fofaviewer.request;
+
+import javafx.scene.control.TableView;
+
+/**
+ * MainController回调，用于在线程中设置调用MainController的方法
+ */
+public interface MainControllerCallback {
+
+    default boolean getFidStatus() {return false;}
+
+    default void queryCall(String queryTxt){}
+
+    default void addSBListener(TableView<?> view){}
+
+    default void setStatusBar(){}
+}
