+++ src/main/java/org/fofaviewer/request/Request.java
@@ -1,0 +1,113 @@
+package org.fofaviewer.request;
+
+import com.alibaba.fastjson.JSON;
+import com.alibaba.fastjson.JSONObject;
+import javafx.animation.PauseTransition;
+import javafx.application.Platform;
+import javafx.collections.FXCollections;
+import javafx.scene.control.*;
+import javafx.scene.layout.BorderPane;
+import javafx.util.Duration;
+import org.controlsfx.control.StatusBar;
+import org.fofaviewer.bean.RequestBean;
+import org.fofaviewer.bean.TabDataBean;
+import org.fofaviewer.bean.TableBean;
+import org.fofaviewer.controls.MyTableView;
+import org.fofaviewer.utils.*;
+import org.tinylog.Logger;
+import java.util.ArrayList;
+import java.util.HashMap;
+import java.util.Map;
+import java.util.Objects;
+import java.util.concurrent.CountDownLatch;
+import java.util.concurrent.Future;
+import java.util.concurrent.Semaphore;
+import java.util.concurrent.atomic.AtomicInteger;
+
+public class Request {
+    private final ArrayList<RequestBean> queryList;
+    private final RequestCallback<Request> callback;
+    private final MainControllerCallback mainControllerCallback;
+    private AtomicInteger succeeded;
+    private Semaphore semaphore; // 设置线程最大数量
+    private CountDownLatch latch;
+    private Future<Void> future;
+
+    public Request(ArrayList<RequestBean> queryList, RequestCallback<Request> callback, MainControllerCallback mainControllerCallback){
+        this.callback = Objects.requireNonNull(callback);
+        this.queryList = queryList;
+        this.mainControllerCallback = mainControllerCallback;
+    }
+
+    @SuppressWarnings("unchecked")
+    public void query(){
+        semaphore = new Semaphore(10);
+        succeeded = new AtomicInteger();
+        latch = new CountDownLatch(queryList.size());
+        future = (Future<Void>) ThreadPoolUtil.submit(() -> {
+            try{
+                launchRequest(queryList, bean -> {
+                    bean.setRequestStatus(RequestStatus.RUNNING);
+                    TabDataBean _tmp = new TabDataBean();
+                    Platform.runLater(() -> this.callback.before(_tmp));
+                    HashMap<String, String> res = RequestUtil.getInstance().getHTML(bean.getRequestUrl(), 10000, 10000);
+                    bean.setResult(res);
+                    if(res.get("code").equals("error")){
+                        bean.setRequestStatus(RequestStatus.FAILED);
+                        Platform.runLater(() -> this.callback.failed(bean));
+                    }else {
+                        bean.setRequestStatus(RequestStatus.SUCCEEDED);
+                        JSONObject obj = JSON.parseObject(bean.getResult().get("msg"));
+                        if (obj.getBoolean("error")) {
+                            DataUtil.showAlert(Alert.AlertType.ERROR, null, obj.getString("errmsg"));
+                            return;
+                        }
+                        if (obj.getInteger("size") < Integer.parseInt(bean.getSize())) {
+                            _tmp.hasMoreData = false;
+                        }
+                        StatusBar bar = new StatusBar();
+                        _tmp.total = obj.getInteger("size");
+                        Label totalLabel = new Label(ResourceBundleUtil.getResource().getString("QUERY_TIPS1") + obj.getString("size"));
+                        Label countLabel = new Label(String.valueOf(obj.getJSONArray("results").size()));
+                        bar.getRightItems().addAll(new ArrayList<Label>() {{
+                            add(totalLabel);
+                            add(new Label(ResourceBundleUtil.getResource().getString("QUERY_TIPS2")));
+                            add(countLabel);
+                            add(new Label(ResourceBundleUtil.getResource().getString("QUERY_TIPS3")));
+                        }});
+                        BorderPane tablePane = new BorderPane();
+                        TableView<TableBean> view = new TableView<>();
+                        Map<String, TableBean> values = (Map<String, TableBean>) DataUtil.loadJsonData(_tmp, obj, null, null, false, view);
+                        view.setItems(FXCollections.observableArrayList(values.values()));
+                        new MyTableView(view, mainControllerCallback);
+                        tablePane.setCenter(view);
+                        tablePane.setBottom(bar);
+                        // 设置延时任务在滚动条界面渲染结束后进行事件绑定
+                        PauseTransition pause = new PauseTransition(Duration.seconds(0.5));
+                        pause.setOnFinished(event -> mainControllerCallback.addSBListener(view));
+                        pause.playFromStart();
+                        Platform.runLater(() -> this.callback.succeeded(tablePane, bar));
+                    }
+                    semaphore.release();
+                    succeeded.incrementAndGet();
+                });
+            }catch (InterruptedException e){
+                Logger.error(e);
+            }
+        });
+    }
+
+    private void launchRequest(ArrayList<RequestBean> queryList, HttpCallback callback) throws InterruptedException {
+        for(RequestBean bean : queryList) {
+            query(bean, callback);
+        }
+        latch.await();
+    }
+
+    private void query(RequestBean bean, HttpCallback callback) throws InterruptedException {
+        semaphore.acquire();
+        callback.onResponse(bean);
+    }
+
+
+}
