+++ src/main/java/org/fofaviewer/utils/HttpCallback.java
@@ -1,0 +1,7 @@
+package org.fofaviewer.utils;
+
+import org.fofaviewer.bean.RequestBean;
+
+public interface HttpCallback {
+    void onResponse(RequestBean bean);
+}
