+++ src/main/resources/locales.properties
@@ -1,69 +1,74 @@
-QUERY_CONTENT=查询条件
-REMOVE_HONEYPOTS=排除干扰
-ABOUT=关于
-HELP=帮助
-SEARCH=查询
-HOMEPAGE=首页
-QUERY_BUTTON=计算并查询
-CERT_LABEL=计算证书序列号
-FAVICON_LABEL=计算favicon hash
-CERT_HINT=请将16进制证书序列号粘贴到此处！
-FAVICON_HINT=请将网站的Favicon的URL粘贴到此处！
-ABOUT_HINT1=有问题请先查看说明
-ABOUT_HINT2=有bug请提交issue
-ERROR_URL=缺少http协议头！
-DIRECTORY_CHOOSER_TITLE=选择导出结果到目录...
-EXPORT_CONFIRM=导出确认
-EXPORT_BUTTON=导出当前查询的所有数据
-EXPORT_HINT1=当前查询条件可查询到
-EXPORT_HINT2=条数据，最多只能导出
-EXPORT_HINT3=条数据，总共
-EXPORT_HINT4=页\n全部导出请直接点击确认，选择导出部分数据请输入页数。
-EXPORT_CONTENT=请输入要导出的页数：
-EXPORT_INPUT_NUM_HINT1=输入的值必须介于0到
-EXPORT_INPUT_NUM_HINT2=之间！
-EXPORT_INPUT_NUM_ERROR=输入的值不是整数！
-EXPORT_TITLE=正在导出数据中...
-EXPORT_HEADER1=数据总量为
-EXPORT_HEADER2=，总共需要导出
-EXPORT_HEADER3=页
-EXPORT_FILENAME=fofa导出结果_
-EXPORT_FILENAME_SHEET1=查询title
-EXPORT_FILENAME_SHEET2=查询结果
+QUERY_CONTENT=\u67E5\u8BE2\u6761\u4EF6
+REMOVE_HONEYPOTS=\u6392\u9664\u5E72\u6270
+ABOUT=\u5173\u4E8E
+HELP=\u5E2E\u52A9
+PROJECT=\u9879\u76EE
+CREATE_PROJECT=\u6253\u5F00\u9879\u76EE
+SAVE_PROJECT=\u4FDD\u5B58\u9879\u76EE
+SEARCH=\u67E5\u8BE2
+HOMEPAGE=\u9996\u9875
+QUERY_BUTTON=\u8BA1\u7B97\u5E76\u67E5\u8BE2
+CERT_LABEL=\u8BA1\u7B97\u8BC1\u4E66\u5E8F\u5217\u53F7
+FAVICON_LABEL=\u8BA1\u7B97favicon hash
+CERT_HINT=\u8BF7\u5C0616\u8FDB\u5236\u8BC1\u4E66\u5E8F\u5217\u53F7\u7C98\u8D34\u5230\u6B64\u5904\uFF01
+FAVICON_HINT=\u8BF7\u5C06\u7F51\u7AD9\u7684Favicon\u7684URL\u7C98\u8D34\u5230\u6B64\u5904\uFF01
+ABOUT_HINT1=\u6709\u95EE\u9898\u8BF7\u5148\u67E5\u770BREADME
+ABOUT_HINT2=\u6709bug\u8BF7\u63D0\u4EA4issue
+ERROR_URL=\u7F3A\u5C11http\u534F\u8BAE\u5934\uFF01
+DIRECTORY_CHOOSER_TITLE=\u9009\u62E9\u5BFC\u51FA\u7ED3\u679C\u5230\u76EE\u5F55...
+EXPORT_CONFIRM=\u5BFC\u51FA\u786E\u8BA4
+EXPORT_BUTTON=\u5BFC\u51FA\u5F53\u524D\u67E5\u8BE2\u7684\u6240\u6709\u6570\u636E
+EXPORT_HINT1=\u5F53\u524D\u67E5\u8BE2\u6761\u4EF6\u53EF\u67E5\u8BE2\u5230
+EXPORT_HINT2=\u6761\u6570\u636E\uFF0C\u6700\u591A\u53EA\u80FD\u5BFC\u51FA
+EXPORT_HINT3=\u6761\u6570\u636E\uFF0C\u603B\u5171
+EXPORT_HINT4=\u9875\n\u5168\u90E8\u5BFC\u51FA\u8BF7\u76F4\u63A5\u70B9\u51FB\u786E\u8BA4\uFF0C\u9009\u62E9\u5BFC\u51FA\u90E8\u5206\u6570\u636E\u8BF7\u8F93\u5165\u9875\u6570\u3002
+EXPORT_CONTENT=\u8BF7\u8F93\u5165\u8981\u5BFC\u51FA\u7684\u9875\u6570\uFF1A
+EXPORT_INPUT_NUM_HINT1=\u8F93\u5165\u7684\u503C\u5FC5\u987B\u4ECB\u4E8E0\u5230
+EXPORT_INPUT_NUM_HINT2=\u4E4B\u95F4\uFF01
+EXPORT_INPUT_NUM_ERROR=\u8F93\u5165\u7684\u503C\u4E0D\u662F\u6574\u6570\uFF01
+EXPORT_TITLE=\u6B63\u5728\u5BFC\u51FA\u6570\u636E\u4E2D...
+EXPORT_HEADER1=\u6570\u636E\u603B\u91CF\u4E3A
+EXPORT_HEADER2=\uFF0C\u603B\u5171\u9700\u8981\u5BFC\u51FA
+EXPORT_HEADER3=\u9875
+EXPORT_FILENAME=fofa\u5BFC\u51FA\u7ED3\u679C_
+EXPORT_FILENAME_SHEET1=\u67E5\u8BE2title
+EXPORT_FILENAME_SHEET2=\u67E5\u8BE2\u7ED3\u679C
 EXPORT_FILENAME_SHEET3=urls
-EXPORT_MESSAGE1=导出成功！文件保存在
-EXPORT_MESSAGE2_1=部分数据导出成功，其中第
-EXPORT_MESSAGE2_2=页加载失败，文件保存在
-EXPORT_ERROR=导出失败！
-LOADDATA_HINT1=已加载
-LOADDATA_HINT2=页
-QUERY_TIPS1=当前查询条件查询到 
-QUERY_TIPS2= 条，当前已加载
-QUERY_TIPS3= 条
-LOAD_CONFIG_ERROR=缺少配置文件config.properties！
-TABLE_HEADER_NO=序号
+EXPORT_MESSAGE1=\u5BFC\u51FA\u6210\u529F\uFF01\u6587\u4EF6\u4FDD\u5B58\u5728
+EXPORT_MESSAGE2_1=\u90E8\u5206\u6570\u636E\u5BFC\u51FA\u6210\u529F\uFF0C\u5176\u4E2D\u7B2C
+EXPORT_MESSAGE2_2=\u9875\u52A0\u8F7D\u5931\u8D25\uFF0C\u6587\u4EF6\u4FDD\u5B58\u5728
+EXPORT_ERROR=\u5BFC\u51FA\u5931\u8D25\uFF01
+LOADDATA_HINT1=\u5DF2\u52A0\u8F7D
+LOADDATA_HINT2=\u9875
+QUERY_TIPS1=\u5F53\u524D\u67E5\u8BE2\u6761\u4EF6\u67E5\u8BE2\u5230 
+QUERY_TIPS2= \u6761\uFF0C\u5F53\u524D\u5DF2\u52A0\u8F7D
+QUERY_TIPS3= \u6761
+LOAD_CONFIG_ERROR=\u7F3A\u5C11\u914D\u7F6E\u6587\u4EF6config.properties\uFF01
+TABLE_HEADER_NO=\u5E8F\u53F7
 TABLE_HEADER_HOST=HOST
-TABLE_HEADER_TITLE=标题
+TABLE_HEADER_TITLE=\u6807\u9898
 TABLE_HEADER_IP=IP
-TABLE_HEADER_PORT=端口
-TABLE_HEADER_DOMAIN=域名
-TABLE_HEADER_PROCOTOL=协议
-TABLE_HEADER_SERVER=Server指纹
+TABLE_HEADER_PORT=\u7AEF\u53E3
+TABLE_HEADER_DOMAIN=\u57DF\u540D
+TABLE_HEADER_PROCOTOL=\u534F\u8BAE
+TABLE_HEADER_SERVER=Server\u6307\u7EB9
 TABLE_HEADER_FID=Fid
-TABLE_HEADER_CERTCN=证书绑定的域名
-TABLE_CONTEXTMENU_COPYLINK=复制链接
-TABLE_CONTEXTMENU_COPYIP=复制IP
-TABLE_CONTEXTMENU_QUERY_IP=查询IP
-TABLE_CONTEXTMENU_QUERY_C-CLASS=查询C段IP
-TABLE_CONTEXTMENU_QUERY_DOAMIN=查询相关域名资产
-TABLE_CONTEXTMENU_QUERY_FAVICON=从fofa搜索favicon相关的资产
-TABLE_CONTEXTMENU_CERT=从fofa搜索证书相关的资产
-TABLE_CONTEXTMENU_FID=从fofa搜索对应Fid的资产
-QUERY_FAVICON_ERROR=该网站未提供favicon
-QUERY_CERT_ERROR=请选中host中带有https的行再点击查询证书相关资产
-QUERY_FID_ERROR=fid不能为空
-CLOSE_OPTIONS=关闭选项
-CLOSE_OTHERS=关闭其他
-CLOSE_ALL=关闭所有
-CLOSE_CHOOSE=关闭选中
+TABLE_HEADER_CERTCN=\u8BC1\u4E66\u7ED1\u5B9A\u7684\u57DF\u540D
+TABLE_CONTEXTMENU_COPYLINK=\u590D\u5236\u94FE\u63A5
+TABLE_CONTEXTMENU_COPYIP=\u590D\u5236IP
+TABLE_CONTEXTMENU_QUERY_IP=\u67E5\u8BE2IP
+TABLE_CONTEXTMENU_QUERY_C-CLASS=\u67E5\u8BE2C\u6BB5IP
+TABLE_CONTEXTMENU_QUERY_DOAMIN=\u67E5\u8BE2\u76F8\u5173\u57DF\u540D\u8D44\u4EA7
+TABLE_CONTEXTMENU_QUERY_FAVICON=\u4ECEfofa\u641C\u7D22favicon\u76F8\u5173\u7684\u8D44\u4EA7
+TABLE_CONTEXTMENU_CERT=\u4ECEfofa\u641C\u7D22\u8BC1\u4E66\u76F8\u5173\u7684\u8D44\u4EA7
+TABLE_CONTEXTMENU_FID=\u4ECEfofa\u641C\u7D22\u5BF9\u5E94Fid\u7684\u8D44\u4EA7
+QUERY_FAVICON_ERROR=\u8BE5\u7F51\u7AD9\u672A\u63D0\u4F9Bfavicon
+QUERY_CERT_ERROR=\u8BF7\u9009\u4E2Dhost\u4E2D\u5E26\u6709https\u7684\u884C\u518D\u70B9\u51FB\u67E5\u8BE2\u8BC1\u4E66\u76F8\u5173\u8D44\u4EA7
+QUERY_FID_ERROR=fid\u4E0D\u80FD\u4E3A\u7A7A
+CLOSE_OPTIONS=\u9009\u9879
+CLOSE_OTHERS=\u5173\u95ED\u5176\u4ED6
+CLOSE_ALL=\u5173\u95ED\u6240\u6709
+CLOSE_CHOOSE=\u5173\u95ED\u9009\u4E2D
+RELOAD_TAB=\u91CD\u65B0\u52A0\u8F7D
 WITH_FID=Fid
+IS_ALL=\u5168\u90E8
