+++ src/main/resources/locales_en_US.properties
@@ -2,6 +2,9 @@ QUERY_CONTENT=content
 REMOVE_HONEYPOTS=Exclude honeypots
 ABOUT=about
 HELP=help
+PORJECT=Project
+CREATE_PROJECT=open project
+SAVE_PROJECT=save project
 SEARCH=search
 HOMEPAGE=Home
 QUERY_BUTTON=handle&query
@@ -65,4 +68,6 @@ CLOSE_OPTIONS=options
 CLOSE_OTHERS=close others
 CLOSE_ALL=close all
 CLOSE_CHOOSE=close choose
+RELOAD_TAB=reload tab
 WITH_FID=Fid
+IS_ALL=all
