+++ src/main/resources/tabpane.css
@@ -1,3 +1,3 @@
 .tab-pane .tab-header-area {
-    -fx-padding: 5 0 0 96;
+    -fx-padding: 5 0 0 76;
 }
