+++ src/main/resources/tinylog.properties
@@ -1,0 +1,11 @@
+writer1=rolling file
+writer1.level=error
+writer1.format={date:yyyy-MM-dd HH:mm:ss.SSS} {level} {file}.java:{line}-{method}:\n{message}
+writer1.file=log/error_{date:yyyy-MM-dd}.log
+writer1.append=true
+writer1.charset=UTF-8
+writer1.writingthread=true
+writer1.policies = startup
+writer2=console
+writer2.level=warn
+writer2.format={date:yyyy-MM-dd HH:mm:ss.SSS} {level} {file}.java:{line}-{method}:\n{message}
