+++ src/test/java/Test.java
@@ -1,5 +1,5 @@
 import org.fofaviewer.bean.FofaBean;
-import org.fofaviewer.utils.RequestHelper;
+import org.fofaviewer.utils.RequestUtil;
 import java.io.FileInputStream;
 import java.io.IOException;
 import java.util.Properties;
@@ -30,12 +30,12 @@ public class Test {
     public static void main(String[] args) throws IOException {
         Properties properties = new Properties();
         properties.load(new FileInputStream("config.properties"));
-        FofaBean client = new FofaBean(properties.getProperty("email").trim(), properties.getProperty("key").trim());
+        FofaBean client = new FofaBean(properties.getProperty("email").trim(), properties.getProperty("key").trim(), properties.getProperty("api"));
         client.setSize(properties.getProperty("maxSize"));
-        RequestHelper helper = RequestHelper.getInstance();
-        String a = "domain=\"baidu.com\"";
+        RequestUtil helper = RequestUtil.getInstance();
+        String a = "致远";
         System.out.println(a);
-        String requestUrl = client.getParam(null, false) + helper.encode(a);
+        String requestUrl = client.getParam(null, false, true) + helper.encode(a);
         System.out.println(requestUrl);
         System.out.println(helper.getHTML(requestUrl, 3000,5000));
 //        Test test = new Test();
